% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.boxcoxar.R
\name{gof.boxcoxar}
\alias{gof.boxcoxar}
\title{get goodness of fit}
\usage{
\method{gof}{boxcoxar}(object, type = "pearson", verbose = FALSE)
}
\arguments{
\item{object}{the output model from fitting a model}

\item{type}{the type of goodness of fit to compute; DIC / LPML / Pearson's residual}

\item{verbose}{FALSE by default; If TRUE, then progress bar will appear}
}
\description{
get goodness of fit
}
