% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{get goodness of fit: DIC / LPML / Pearson's residual}
\usage{
gof(object, type = "pearson", verbose = FALSE)
}
\arguments{
\item{object}{the output model from fitting a model}

\item{type}{the type of goodness of fit to compute; DIC / LPML / Pearson's residual}

\item{verbose}{FALSE by default; If TRUE, then progress bar will appear}
}
\value{
dataframe containing the goodness of fit measure
}
\description{
get goodness of fit: DIC / LPML / Pearson's residual
}
