% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvboxcoxar.R
\name{mvboxcoxar}
\alias{mvboxcoxar}
\title{Fit Bayesian Multivariate Box-Cox Transformed AR(1)}
\usage{
mvboxcoxar(yobs, miss, a, xobs, mcmc = list(), prior = list(),
  verbose = FALSE)
}
\arguments{
\item{yobs}{a matrix of measured energy usage to be used as the response variable}

\item{miss}{a integer matrix of binary values (0 or 1), each element of which indicates whether the response variable is missing}

\item{a}{the shift parameter for the Box-Cox. By default, set to the minimum value minus one.}

\item{xobs}{the covariate matrix for the systemic equations for both the regression mean and Box-Cox lambda}

\item{mcmc}{list of MCMC-related parameters: number of burn-ins (ndiscard), number of thinning(nskip), and posterior sample size (nkeep)}

\item{prior}{list of hyperparameters; when not given, algorithm will run in default setting}

\item{verbose}{logical variable for printing progress bar. Default to FALSE.}
}
\value{
a dataframe with input arguments, posterior samples, Metropolis algorithm acceptance rates, etc
}
\description{
This is a function that fits a univariate Box-Cox model for energy usage
}
\examples{
\dontrun{
data(df)
library(magrittr)
library(dplyr)
library(readr)
library(lubridate)
x <- df \%>\% filter(TimeStamp >= ymd_hms("2018-07-01 00:00:00"),
                   TimeStamp < ymd_hms("2018-08-01 00:00:00"))
y <- x$Reading \%>\% scale(center=FALSE) \%>\% drop
n <- length(y)
x1 <- x$TempRead \%>\% scale \%>\% drop
x2 <- x$HumidityRead \%>\% scale \%>\% drop
x3 <- 1*(hour(x$TimeStamp) \%in\% c(0:6, 19:23))

xobs <- cbind(1, x1, x2, x3)
zobs <- cbind(1, x1, x2)
uobs <- cbind(1, x1, x2, x3)

fit <- mvboxcoxar(y, integer(length(y)), mcmc=list(ndiscard=10000, nkeep = 20000),
         xobs = xobs, zobs = zobs, uobs = uobs, verbose=TRUE)
}
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
