% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_monitor.R
\name{single_monitor}
\alias{single_monitor}
\title{Testing whether a given year is out-of-control}
\usage{
single_monitor(y_surveill, yh_bar, sigma, nyears, sig_level, C, B1, B2, B3,
  verbose = FALSE)
}
\arguments{
\item{y_surveill}{a vector of measurements to monitor}

\item{yh_bar}{a vector of historical measurements to estimate the parameters}

\item{sigma}{the estimate of the standard deviation}

\item{nyears}{# of years for the historical data}

\item{sig_level}{significance level}

\item{C}{the constant C that determines the power}

\item{B1}{# of outer loop iterations}

\item{B2}{# of middle loop iterations}

\item{B3}{# of innermost loop iterations}

\item{verbose}{logical variable for printing progress bar. Default to FALSE.}
}
\value{
either 0 or 1 that indicates whether the null hypothesis has been rejected or not
}
\description{
This function performs hypothesis testing for monitoring whether a given year is
out-of-control or in-control
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
