\name{TSEnergy-package}
\alias{TSEnergy-package}
\alias{TSEnergy}
\docType{package}
\title{
\packageTitle{TSEnergy}
}
\description{
\packageDescription{TSEnergy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TSEnergy}
\packageIndices{TSEnergy}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{TSEnergy}

Maintainer: \packageMaintainer{TSEnergy}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
